Add jsmediatags library to parse ID3 tags
This library adds ~15kb gzipped (52kb unzipped) to the page weight.

jsmediatags is licensed under the BSD 3-clause license, I've prepended
a copy of the license to the minified library. It seems to be the
easiest way to be compliant.

https://github.com/aadsm/jsmediatags/blob/master/LICENSE.md

In terms of the implementation, I wrap their async API in a promise
so that it matches style with get_audio_duration which is the closest
analogue I could find in the source.

If we fail to read tags from a file, just silently fallback to the
file.name